requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.2.0
requests==2.31.0
ipython==8.16.1